requirements.txt
python-dotenv
amadeus
mysql-connector-python
pytz
